requirements.test.txt
# linters such as flake8 and pylint should be pinned, as new releases
# make new things fail. Manually update these pins when pulling in a
# new version

aioresponses==0.7.2
algoliasearch==2.6.2
codecov==2.1.12
coverage>=6.4.4
mypy==0.991
pytest>=7.1.0
pytest-cov>=3.0.0
pytest-mock>=3.10
pytest-homeassistant-custom-component==0.12.20
typing-extensions==4.4.0
